backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pytest==7.4.2
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.7 